Add Moondream package

+++ ros1/catkin_ws/devel/include/limo_application/image_processingConfig.h
@@ -290,17 +290,17 @@ int yellow_s_high;
 
 
 
-//#line 25 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 25 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       int yellow_h_low;
-//#line 26 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 26 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       int yellow_h_high;
-//#line 27 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 27 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       int yellow_l_low;
-//#line 28 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 28 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       int yellow_l_high;
-//#line 29 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 29 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       int yellow_s_low;
-//#line 30 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 30 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       int yellow_s_high;
 //#line 231 "/opt/ros/noetic/share/dynamic_reconfigure/cmake/../templates/ConfigType.h.template"
 
@@ -328,16 +328,16 @@ int yellow_s_high;
       {
         ROS_ERROR("image_processingConfig::__fromMessage__ called with an unexpected parameter.");
         ROS_ERROR("Booleans:");
-        for (unsigned int i = 0; i < msg.bools.size(); i++)
+        for (size_t i = 0; i < msg.bools.size(); i++)
           ROS_ERROR("  %s", msg.bools[i].name.c_str());
         ROS_ERROR("Integers:");
-        for (unsigned int i = 0; i < msg.ints.size(); i++)
+        for (size_t i = 0; i < msg.ints.size(); i++)
           ROS_ERROR("  %s", msg.ints[i].name.c_str());
         ROS_ERROR("Doubles:");
-        for (unsigned int i = 0; i < msg.doubles.size(); i++)
+        for (size_t i = 0; i < msg.doubles.size(); i++)
           ROS_ERROR("  %s", msg.doubles[i].name.c_str());
         ROS_ERROR("Strings:");
-        for (unsigned int i = 0; i < msg.strs.size(); i++)
+        for (size_t i = 0; i < msg.strs.size(); i++)
           ROS_ERROR("  %s", msg.strs[i].name.c_str());
         // @todo Check that there are no duplicates. Make this error more
         // explicit.
@@ -441,75 +441,75 @@ int yellow_s_high;
     {
 image_processingConfig::GroupDescription<image_processingConfig::DEFAULT, image_processingConfig> Default("Default", "", 0, 0, true, &image_processingConfig::groups);
 image_processingConfig::GroupDescription<image_processingConfig::DEFAULT::YELLOW_LANE_DETECTION, image_processingConfig::DEFAULT> Yellow_Lane_Detection("Yellow_Lane_Detection", "", 0, 1, true, &image_processingConfig::DEFAULT::yellow_lane_detection);
-//#line 25 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 25 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __min__.yellow_h_low = 0;
-//#line 25 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 25 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __max__.yellow_h_low = 255;
-//#line 25 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 25 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __default__.yellow_h_low = 50;
-//#line 25 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 25 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       Yellow_Lane_Detection.abstract_parameters.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_h_low", "int", 4, "H Channel Low threshold", "", &image_processingConfig::yellow_h_low)));
-//#line 25 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 25 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __param_descriptions__.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_h_low", "int", 4, "H Channel Low threshold", "", &image_processingConfig::yellow_h_low)));
-//#line 26 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 26 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __min__.yellow_h_high = 0;
-//#line 26 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 26 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __max__.yellow_h_high = 255;
-//#line 26 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 26 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __default__.yellow_h_high = 50;
-//#line 26 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 26 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       Yellow_Lane_Detection.abstract_parameters.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_h_high", "int", 5, "H Channel High threshold", "", &image_processingConfig::yellow_h_high)));
-//#line 26 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 26 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __param_descriptions__.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_h_high", "int", 5, "H Channel High threshold", "", &image_processingConfig::yellow_h_high)));
-//#line 27 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 27 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __min__.yellow_l_low = 0;
-//#line 27 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 27 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __max__.yellow_l_low = 255;
-//#line 27 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 27 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __default__.yellow_l_low = 50;
-//#line 27 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 27 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       Yellow_Lane_Detection.abstract_parameters.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_l_low", "int", 6, "L ChannelLow threshold", "", &image_processingConfig::yellow_l_low)));
-//#line 27 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 27 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __param_descriptions__.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_l_low", "int", 6, "L ChannelLow threshold", "", &image_processingConfig::yellow_l_low)));
-//#line 28 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 28 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __min__.yellow_l_high = 0;
-//#line 28 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 28 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __max__.yellow_l_high = 255;
-//#line 28 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 28 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __default__.yellow_l_high = 50;
-//#line 28 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 28 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       Yellow_Lane_Detection.abstract_parameters.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_l_high", "int", 7, "L Channel High threshold", "", &image_processingConfig::yellow_l_high)));
-//#line 28 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 28 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __param_descriptions__.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_l_high", "int", 7, "L Channel High threshold", "", &image_processingConfig::yellow_l_high)));
-//#line 29 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 29 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __min__.yellow_s_low = 0;
-//#line 29 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 29 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __max__.yellow_s_low = 255;
-//#line 29 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 29 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __default__.yellow_s_low = 50;
-//#line 29 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 29 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       Yellow_Lane_Detection.abstract_parameters.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_s_low", "int", 8, "S Channel Low threshold", "", &image_processingConfig::yellow_s_low)));
-//#line 29 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 29 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __param_descriptions__.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_s_low", "int", 8, "S Channel Low threshold", "", &image_processingConfig::yellow_s_low)));
-//#line 30 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 30 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __min__.yellow_s_high = 0;
-//#line 30 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 30 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __max__.yellow_s_high = 255;
-//#line 30 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 30 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __default__.yellow_s_high = 50;
-//#line 30 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 30 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       Yellow_Lane_Detection.abstract_parameters.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_s_high", "int", 9, "S Channel High threshold", "", &image_processingConfig::yellow_s_high)));
-//#line 30 "/home/wego/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
+//#line 30 "/root/adi_ws/ros1/catkin_ws/src/limo_application-master/cfg/lane_detection/image_processing.cfg"
       __param_descriptions__.push_back(image_processingConfig::AbstractParamDescriptionConstPtr(new image_processingConfig::ParamDescription<int>("yellow_s_high", "int", 9, "S Channel High threshold", "", &image_processingConfig::yellow_s_high)));
-//#line 124 "/opt/ros/noetic/lib/python3/dist-packages/dynamic_reconfigure/parameter_generator_catkin.py"
+//#line 125 "/opt/ros/noetic/lib/python3/dist-packages/dynamic_reconfigure/parameter_generator_catkin.py"
       Yellow_Lane_Detection.convertParams();
-//#line 124 "/opt/ros/noetic/lib/python3/dist-packages/dynamic_reconfigure/parameter_generator_catkin.py"
+//#line 125 "/opt/ros/noetic/lib/python3/dist-packages/dynamic_reconfigure/parameter_generator_catkin.py"
       Default.groups.push_back(image_processingConfig::AbstractGroupDescriptionConstPtr(new image_processingConfig::GroupDescription<image_processingConfig::DEFAULT::YELLOW_LANE_DETECTION, image_processingConfig::DEFAULT>(Yellow_Lane_Detection)));
-//#line 124 "/opt/ros/noetic/lib/python3/dist-packages/dynamic_reconfigure/parameter_generator_catkin.py"
+//#line 125 "/opt/ros/noetic/lib/python3/dist-packages/dynamic_reconfigure/parameter_generator_catkin.py"
       __group_descriptions__.push_back(image_processingConfig::AbstractGroupDescriptionConstPtr(new image_processingConfig::GroupDescription<image_processingConfig::DEFAULT::YELLOW_LANE_DETECTION, image_processingConfig::DEFAULT>(Yellow_Lane_Detection)));
-//#line 246 "/opt/ros/noetic/lib/python3/dist-packages/dynamic_reconfigure/parameter_generator_catkin.py"
+//#line 247 "/opt/ros/noetic/lib/python3/dist-packages/dynamic_reconfigure/parameter_generator_catkin.py"
       Default.convertParams();
-//#line 246 "/opt/ros/noetic/lib/python3/dist-packages/dynamic_reconfigure/parameter_generator_catkin.py"
+//#line 247 "/opt/ros/noetic/lib/python3/dist-packages/dynamic_reconfigure/parameter_generator_catkin.py"
       __group_descriptions__.push_back(image_processingConfig::AbstractGroupDescriptionConstPtr(new image_processingConfig::GroupDescription<image_processingConfig::DEFAULT, image_processingConfig>(Default)));
 //#line 369 "/opt/ros/noetic/share/dynamic_reconfigure/cmake/../templates/ConfigType.h.template"
